pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kres"
version = "1.0.0"
description = "Kubernetes Resource Restarter"
authors = [
    { name = "Shashank Palla", email = "shashankpalla44@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pyyaml>=6.0.1",
    "requests==2.32.4",
    "cryptography==45.0.5",
    "fastapi==0.116.1",
    "uvicorn==0.35.0"
]

[project.scripts]
kres = "kres.main:main"